backend/requirements.txt
annotated-types==0.6.0
anyio==4.0.0
assemblyai==0.18.0
blinker==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
ffmpeg==1.4
ffmpeg-python==0.2.0
Flask==3.0.0
Flask-Cors==4.0.0
future==0.18.3
google-api-core==2.12.0
google-auth==2.23.2
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.14.2
googleapis-common-protos==1.60.0
grpcio==1.59.0
grpcio-status==1.59.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
proto-plus==1.22.3
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.4.2
pydantic_core==2.10.1
pydub==0.25.1
pyngrok==7.0.0
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
sniffio==1.3.0
typing_extensions==4.8.0
urllib3==2.0.6
websockets==11.0.3
Werkzeug==3.0.0
